Authorization with Gate

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,7 @@ use App\Post;
 use App\Like;
 use App\Tag;
 use Auth;
+use Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Session\Store;
@@ -110,6 +111,9 @@ class PostController extends Controller
         ]);
         
         $post = Post::find($request->input('id'));
+        if(Gate::denies('manipulate-post', $post)){
+            return redirect()->back();
+        }
         $post->title = $request->input('title');
         $post->content = $request->input('content');
         $post->save();
@@ -128,6 +132,9 @@ class PostController extends Controller
             return redirect()->back();
         }
         $post = Post::find($id);
+        if(Gate::denies('manipulate-post', $post)){
+            return redirect()->back();
+        }
         $post->likes()->delete();
         $post->tags()->detach();
         $post->delete();
